Made conditions and effects data preparation code shared. Reused in PC and NPC sheet data prep.
